+++ activation/validation.go
@@ -132,7 +132,7 @@ func validateMerkleProof(leaf []byte, proof *types.MerkleProof, expectedRoot []b
 			hexNodes = append(hexNodes, n.Hex())
 		}
 		return fmt.Errorf(
-			"invalid merkle proof, calculated root does not match the proof root, leaf: %v, nodes: %v, expected root: %v",
+			"invalid merkle proof, calculated root does not match proof root, leaf: %v, nodes: %v, expected root: %v",
 			util.Encode(leaf),
 			hexNodes,
 			util.Encode(expectedRoot),
@@ -223,7 +223,8 @@ func (v *Validator) VRFNonce(
 		LabelsPerUnit:   PostMetadata.LabelsPerUnit,
 	}
 
-	if err := verifying.VerifyVRFNonce((*uint64)(vrfNonce), meta, verifying.WithLabelScryptParams(v.scrypt)); err != nil {
+	err := verifying.VerifyVRFNonce((*uint64)(vrfNonce), meta, verifying.WithLabelScryptParams(v.scrypt))
+	if err != nil {
 		return fmt.Errorf("verify VRF nonce: %w", err)
 	}
 	return nil
@@ -275,7 +276,7 @@ func (*Validator) NIPostChallenge(challenge *types.NIPostChallenge, atxs atxProv
 	}
 
 	if prevATX.Sequence+1 != challenge.Sequence {
-		return fmt.Errorf("sequence number is not one more than prev sequence number")
+		return errors.New("sequence number is not one more than prev sequence number")
 	}
 	return nil
 }