+++ api/grpcserver/v2alpha1/v2alpha1_test.go
@@ -3,6 +3,7 @@ package v2alpha1
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -11,20 +12,28 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/spacemeshos/go-spacemesh/api/grpcserver"
+	"github.com/spacemeshos/go-spacemesh/common/types"
 )
+
+const (
+	genTimeUnix   = 1000000
+	layerDuration = 10 * time.Second
+)
+
+var genesisID = types.Hash20{}
 
 func launchServer(tb testing.TB, services ...grpcserver.ServiceAPI) (grpcserver.Config, func()) {
 	cfg := grpcserver.DefaultTestConfig()
-	grpcService, err := grpcserver.NewWithServices(cfg.PublicListener, zaptest.NewLogger(tb).Named("grpc"), cfg, services)
+	grpc, err := grpcserver.NewWithServices(cfg.PublicListener, zaptest.NewLogger(tb).Named("grpc"), cfg, services)
 	require.NoError(tb, err)
 
 	// start gRPC server
-	require.NoError(tb, grpcService.Start())
+	require.NoError(tb, grpc.Start())
 
 	// update config with bound addresses
-	cfg.PublicListener = grpcService.BoundAddress
+	cfg.PublicListener = grpc.BoundAddress
 
-	return cfg, func() { assert.NoError(tb, grpcService.Close()) }
+	return cfg, func() { assert.NoError(tb, grpc.Close()) }
 }
 
 func dialGrpc(ctx context.Context, tb testing.TB, cfg grpcserver.Config) *grpc.ClientConn {