+++ atxsdata/warmup.go
@@ -11,20 +11,20 @@ import (
 	"github.com/spacemeshos/go-spacemesh/sql/layers"
 )
 
-func Warm(db *sql.Database, opts ...Opt) (*Data, error) {
-	cache := New(opts...)
+func Warm(db *sql.Database, keep types.EpochID) (*Data, error) {
+	cache := New()
 	tx, err := db.Tx(context.Background())
 	if err != nil {
 		return nil, err
 	}
 	defer tx.Release()
-	if err := Warmup(tx, cache); err != nil {
+	if err := Warmup(tx, cache, keep); err != nil {
 		return nil, fmt.Errorf("warmup %w", err)
 	}
 	return cache, nil
 }
 
-func Warmup(db sql.Executor, cache *Data) error {
+func Warmup(db sql.Executor, cache *Data, keep types.EpochID) error {
 	latest, err := atxs.LatestEpoch(db)
 	if err != nil {
 		return err
@@ -33,7 +33,11 @@ func Warmup(db sql.Executor, cache *Data) error {
 	if err != nil {
 		return err
 	}
-	cache.OnEpoch(applied.GetEpoch())
+	var evict types.EpochID
+	if applied.GetEpoch() > keep {
+		evict = applied.GetEpoch() - keep - 1
+	}
+	cache.EvictEpoch(evict)
 
 	var ierr error
 	if err := atxs.IterateAtxsData(db, cache.Evicted(), latest,