+++ node/node.go
@@ -81,6 +81,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/syncer"
 	"github.com/spacemeshos/go-spacemesh/syncer/atxsync"
 	"github.com/spacemeshos/go-spacemesh/syncer/blockssync"
+	"github.com/spacemeshos/go-spacemesh/syncer/malsync"
 	"github.com/spacemeshos/go-spacemesh/system"
 	"github.com/spacemeshos/go-spacemesh/timesync"
 	timeCfg "github.com/spacemeshos/go-spacemesh/timesync/config"
@@ -452,7 +453,7 @@ func (app *App) LoadCheckpoint(ctx context.Context) (*checkpoint.PreservedData, 
 		return nil, nil
 	}
 	if restore == 0 {
-		return nil, fmt.Errorf("restore layer not set")
+		return nil, errors.New("restore layer not set")
 	}
 	nodeIDs := make([]types.NodeID, len(app.signers))
 	for i, sig := range app.signers {
@@ -532,7 +533,7 @@ func (app *App) Initialize() error {
 			app.log.Error("genesis config updated after node initialization, if this update is required delete config"+
 				" at %s.\ndiff:\n%s", gpath, diff,
 			)
-			return fmt.Errorf("genesis config updated after node initialization")
+			return errors.New("genesis config updated after node initialization")
 		}
 	}
 
@@ -781,7 +782,8 @@ func (app *App) initServices(ctx context.Context) error {
 
 	if app.Config.HareEligibility.ConfidenceParam >= app.Config.BaseConfig.LayersPerEpoch {
 		return fmt.Errorf(
-			"confidence param should be smaller than layers per epoch. eligibility-confidence-param: %d. layers-per-epoch: %d",
+			"confidence param should be smaller than layers per epoch. eligibility-confidence-param: %d. "+
+				"layers-per-epoch: %d",
 			app.Config.HareEligibility.ConfidenceParam,
 			app.Config.BaseConfig.LayersPerEpoch,
 		)
@@ -861,6 +863,9 @@ func (app *App) initServices(ctx context.Context) error {
 	syncerConf.SyncCertDistance = app.Config.Tortoise.Hdist
 	syncerConf.Standalone = app.Config.Standalone
 
+	if app.Config.P2P.MinPeers < app.Config.Sync.MalSync.MinSyncPeers {
+		app.Config.Sync.MalSync.MinSyncPeers = max(1, app.Config.P2P.MinPeers)
+	}
 	app.syncLogger = app.addLogger(SyncLogger, lg)
 	newSyncer := syncer.NewSyncer(
 		app.cachedDB,
@@ -875,6 +880,11 @@ func (app *App) initServices(ctx context.Context) error {
 			atxsync.WithConfig(app.Config.Sync.AtxSync),
 			atxsync.WithLogger(app.syncLogger.Zap()),
 		),
+		malsync.New(fetcher, app.db, app.localDB,
+			malsync.WithConfig(app.Config.Sync.MalSync),
+			malsync.WithLogger(app.syncLogger.Zap()),
+			malsync.WithPeerErrMetric(syncer.MalPeerError),
+		),
 		syncer.WithConfig(syncerConf),
 		syncer.WithLogger(app.syncLogger),
 	)
@@ -882,7 +892,8 @@ func (app *App) initServices(ctx context.Context) error {
 	beaconProtocol.SetSyncState(newSyncer)
 	app.hOracle.SetSync(newSyncer)
 
-	if err := app.Config.HARE3.Validate(time.Duration(app.Config.Tortoise.Zdist) * app.Config.LayerDuration); err != nil {
+	err = app.Config.HARE3.Validate(time.Duration(app.Config.Tortoise.Zdist) * app.Config.LayerDuration)
+	if err != nil {
 		return err
 	}
 	logger := app.addLogger(HareLogger, lg).Zap()
@@ -1458,6 +1469,13 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		service := v2alpha1.NewRewardStreamService(app.db)
 		app.grpcServices[svc] = service
 		return service, nil
+	case v2alpha1.Network:
+		service := v2alpha1.NewNetworkService(
+			app.clock.GenesisTime(),
+			app.Config.Genesis.GenesisID(),
+			app.Config.LayerDuration)
+		app.grpcServices[svc] = service
+		return service, nil
 	}
 	return nil, fmt.Errorf("unknown service %s", svc)
 }
@@ -1588,10 +1606,10 @@ func (app *App) startAPIServices(ctx context.Context) error {
 		svc.(*grpcserver.SmesherService).SetPostServiceConfig(app.Config.POSTService)
 		if app.Config.SMESHING.Start {
 			if app.Config.SMESHING.CoinbaseAccount == "" {
-				return fmt.Errorf("smeshing enabled but no coinbase account provided")
+				return errors.New("smeshing enabled but no coinbase account provided")
 			}
 			if len(app.signers) > 1 {
-				return fmt.Errorf("supervised smeshing cannot be started in a multi-smeshing setup")
+				return errors.New("supervised smeshing cannot be started in a multi-smeshing setup")
 			}
 			if err := app.postSupervisor.Start(
 				app.Config.POSTService,
@@ -1619,7 +1637,7 @@ func (app *App) startAPIServices(ctx context.Context) error {
 
 	if len(app.Config.API.JSONListener) > 0 {
 		if len(publicSvcs) == 0 {
-			return fmt.Errorf("start json server without public services")
+			return errors.New("start json server without public services")
 		}
 		app.jsonAPIServer = grpcserver.NewJSONHTTPServer(
 			app.Config.API.JSONListener,
@@ -1785,10 +1803,14 @@ func (app *App) setupDBs(ctx context.Context, lg log.Log) error {
 		)
 	}
 	app.log.Info("starting cache warmup")
+	applied, err := layers.GetLastApplied(app.db)
+	if err != nil {
+		return err
+	}
 	start := time.Now()
 	data, err := atxsdata.Warm(
 		app.db,
-		atxsdata.WithCapacityFromLayers(app.Config.Tortoise.WindowSize, app.Config.LayersPerEpoch),
+		app.Config.Tortoise.WindowSizeEpochs(applied),
 	)
 	if err != nil {
 		return err